% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunc.R
\name{DGP}
\alias{DGP}
\title{Data generating process}
\usage{
DGP(beta, n, p, bOut = FALSE)
}
\arguments{
\item{beta}{a binary regression parameter vector.}

\item{n}{a number of rows of the design matrix X}

\item{p}{a number of columns of the design matrix X}

\item{bOut}{a logical value for generating outliers in the design matrix X. Default is FALSE.}
}
\value{
A list of the following values:
\describe{
\item{DM}{ a data frame that includes the binary response Y and design matrix X}

\item{Y}{the binary response Y}

\item{X}{the design matrix X}

}
}
\description{
Generate a data frame including a binary response and design matrix
}
\examples{
####################
n=20
beta = c(1,-2,3.5)
p=3

###### Generate a data frame including a response and design matrix

lst = DGP(beta, n, p, bOut=FALSE)

DM = lst$DM    ### a data frame
Y = lst$Y      ### a binary response  
X = lst$X      ### a design matrix

link = "Logit"
lst = jwglm("Y~Intercept+X1+X2-1", data=DM, nIter=1000)

}
\seealso{
jwglm()
}
