% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwglm-package.R
\docType{package}
\name{jwglm-package}
\alias{jwglm-package}
\title{jwglm: Binary regression estimation using an L2 type loss function}
\description{
This package proposes a novel method to estimate parameters in a binary regression model whose link function is a logit, probit, or loglog function.
}
\keyword{internal}
